+++ package.json
@@ -6,6 +6,7 @@
     "cra-template": "1.2.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1"
   },
   "scripts": {
