+++ src/components/Header.css
@@ -8,9 +8,9 @@
     display: flex;
     align-items: center;
   }
-  .navbar-brand a:link, .navbar-brand a:hover{color:#f9c10b !important;}
+  .navbar-brand a:link, .navbar-brand a:hover{color:#06A !important;}
   .notification-top-bar {
-    background: #f9c10b;
+    background: #06A;
     padding-top: 10px;
     padding-bottom: 10px;
     color:#fff;
@@ -49,7 +49,7 @@
   .site-header .navbar-brand {
     font-size: 32px;
     font-weight:700;
-    color:#f9c10b;
+    color:#06A;
     text-decoration: none;
 }
 .navbar-brand {
@@ -63,10 +63,10 @@
 .navbar-brand {
   font-size: 32px;
   font-weight: 700;
-    color: #f9c10b;
+    color: #06A;
 }
 .navbar-nav a:link{
-  color: #717275;
+  color: #06a;
 }
 .navbar-nav li{
   text-align: center;
