+++ src/components/aboutus.js
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Header.css"; // Import the CSS file
+
+const aboutus = () => {
+  return (
+    <>about us page</>
+  )
+}
+export default aboutus;
