+++ src/components/contactus.js
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Header.css"; // Import the CSS file
+
+const contactus = () => {
+  return (
+    <>contactus page</>
+  )
+}
+export default contactus;
