+++ src/components/notFound.js
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Header.css"; // Import the CSS file
+
+const notFound = () => {
+  return (
+    <>not found page1</>
+  )
+}
+export default notFound;
